requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.3.1
billiard==3.6.4.0
cached-property==1.5.2
cachetools==4.2.1
celery==4.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.3
django-appconf==1.0.4
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-coreapi==2.3.0
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-environ==0.4.5
django-filter==2.4.0
django-googledrive-storage==1.5.2
django-oauth-toolkit==1.4.1
django-phonenumber-field==5.0.0
django-smtp-ssl==1.0
django-storages==1.11.1
django-timezone-field==4.2.1
djangorestframework==3.12.2
emoji==0.6.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.0
idna==2.10
importlib-metadata==4.0.1
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
ndg-httpsclient==0.5.1
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pandas==1.1.4
phonenumbers==8.12.20
prompt-toolkit==3.0.19
protobuf==3.15.5
psycopg2==2.9.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
redis==3.2.1
requests==2.25.1
rsa==4.7.2
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.1
